hardening subsuming load checks

+++ src/hotspot/share/opto/gcm.cpp
@@ -1580,7 +1580,7 @@ void PhaseCFG::global_code_motion() {
   {
     Compile::TracePhase tp("localcodemotion", &timers[_t_localcodemotion]);
     PhaseLCM lcm(*this, regalloc);
-    for (size_t i = 0; i < number_of_blocks(); i++) {
+    for (uint i = 0; i < number_of_blocks(); i++) {
       bool succeeded = lcm.schedule(*get_block(i));
       if (!succeeded) {
         _regalloc = nullptr;
